requirements.txt
torch
torchvision
transformers
pillow
mss
easygui
